+++ CHANGELOG.md
@@ -2,13 +2,14 @@
 All notable version changes would be documented in this file.
 
 ## 0.1.x
-> Minor version released at 2020.03.29
+> Minor version released at 2020.03.29 ~ 2021.02.05
 - Published initial build
 - Chores
   - Update dependencies
   - Add testcase files
   - Refined all source codes with usage of @mornya/vessel v4
   - Runs code linting and testing with @lintest/cli
+- Performance Improvements: Errors and warnings messages were tuned
 - Bugs fixed
   - Error identifying assets
   - GZipped-size is not displayed when running webpack-dev-server